Refactor: Move settings and GUI related code to a separate module
This commit moves the settings and GUI-related functionality into a new module named `gui.py`. This improves the organization and separation of concerns within the project, making it easier to maintain and understand.

The following changes were made:

- The `settings.py` module was removed, and its contents were moved to `gui.py`.
- The `utils/sublime.py` module was renamed to `gui.py`.
- The `api.py`, `models.py`, and `operations.py` modules were updated to import the necessary functions from `gui.py`.

This refactoring makes the codebase more modular and easier to understand. It also ensures that the GUI-related logic is separated from the core application logic.
